若干细节修改；流式加载小优化

+++ Nix/Nix/NXPassMaterial.cpp
@@ -691,8 +691,8 @@ void NXReadbackPassMaterial::ReadbackTexture(ID3D12GraphicsCommandList* pCmdList
 	if (NXReadbackSys->BuildTask(pTexture, ctx))
 	{
 		// 纹理和buffer不太一样，需要根据对纹理抓取的结果 分配实际回读的内存大小
-		uint32_t stride = ctx.rowSizeInBytes / pTexture->GetWidth();
-		uint32_t arraySize = ctx.numRows * ctx.rowSizeInBytes / stride;
+		uint32_t stride = (uint32_t)ctx.rowSizeInBytes / pTexture->GetWidth();
+		uint32_t arraySize = ctx.numRows * (uint32_t)ctx.rowSizeInBytes / stride;
 		if (m_pOutData->GetStride() != stride || m_pOutData->GetByteSize() != stride * arraySize)
 			m_pOutData->Create(stride, arraySize);
